+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.3.0] - 2020-05-19
+
 - Added a `konverter-vault show` command that supports YAML and JSON output
 - Added support for `terraform` output format in `konverter-vault show` command
   that allows using it as an [external data source in
@@ -20,7 +22,8 @@
 
 Initial release
 
-[unreleased]: https://github.com/westphahl/konverter/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/westphahl/konverter/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/westphahl/konverter/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/westphahl/konverter/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/westphahl/konverter/compare/v0.1.0...v0.1.1
 [v0.1.0]: https://github.com/westphahl/konverter/releases/tag/v0.1.0